CDM_UI_Backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
neo4j>=5.15.0
python-dotenv>=1.0.0
pydantic>=2.5.0
python-multipart>=0.0.6
httpx>=0.25.0
